+++ server-actions-for-next-pages/src/turbopackLoader.ts
@@ -3,7 +3,7 @@ import { transform } from '@babel/core';
 import { plugins } from '.';
 import { logger } from './utils';
 
-export default async function (
+module.exports = async function (
   this: LoaderThis<any>,
   source: string,
   map: any,
@@ -13,6 +13,11 @@ export default async function (
     map = JSON.parse(map);
   }
   try {
+    // Skip transformation if file doesn't have the directive
+    if (!source.includes('"poor man\'s use server"') && !source.includes("'poor man's use server'")) {
+      return callback(null, source, map);
+    }
+    
     const options = this.getOptions();
 
     const res = transform(source || '', {
@@ -40,6 +45,8 @@ export default async function (
     callback(e);
   }
 }
+
+module.exports.default = module.exports;
 
 export type LoaderThis<Options> = {
   /**